Stub out app arch
