+++ server/constants/index.js
@@ -1,0 +1,5 @@
+const neoPixelConstants = require('./neo-pixel');
+
+module.exports = {
+  ...neoPixelConstants
+};
