+++ server/constants/neo-pixel.js
@@ -1,0 +1,1 @@
+module.exports.NUM_LEDS = 200;
